requirements.txt
openai>=1.3.0
anthropic>=0.7.0
gitpython>=3.1.40
radon>=6.0.1
lizard>=1.17.10
networkx>=3.2.1
matplotlib>=3.8.0
numpy>=1.26.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pyyaml>=6.0.1
colorlog>=6.7.0
tqdm>=4.66.1
